m6_practice/requirements.txt
torch==2.7.0
torchvision==0.22.0
typer==0.15.2
fastapi==0.115.12
uvicorn==0.34.2
invoke==2.2.0
matplotlib==3.10.0
hydra-core==1.3.2
wandb==0.19.11
loguru==0.7.3
python-dotenv==1.1.0
scikit-learn==1.6.1
lightning==2.5.1
pytest==8.3.5